+++ cloneHD.cpp
@@ -131,10 +131,12 @@ int main (int argc, const char * argv[]){
   myClone.baf_pen  = (opts.baf_pen > 0.0)  ? opts.baf_pen : 1.0;// BAF penalty for complex chr status
   myClone.snv_fpr  = (opts.snv_fpr > 0.0)  ? opts.snv_fpr : 1.0e-4;//SNV false-positive rate
   myClone.snv_err  = (opts.snv_err >= 0.0) ? opts.snv_err : 0.0;   //SNV frequency of false positives
-  myClone.snv_pen  = (opts.snv_pen > 0.0)  ? opts.snv_pen : (bafEmit.is_set ? 0.01 : 0.5);//penalty for SNV genotypes higher than max
+  myClone.snv_pen  = (opts.snv_pen > 0.0)  ? opts.snv_pen : (bafEmit.is_set ? 0.01 : 0.5);//penalty for high SNV genotypes
   myClone.bulk_fix = opts.bulk_fix;
-  //mask for maximum total c.n. per chromosome or genomwide
-  myClone.get_maxcn_mask( opts.maxcn_mask_fn,  opts.maxcn);
+  // *** GET MAXCN MASK ***
+  get_maxcn_mask( opts.maxcn_mask_fn, &myClone, opts.maxcn);
+  if (myClone.maxcn > opts.maxcn)
+    printf("Found new higher maximum copy number %i in %s.\n", myClone.maxcn, opts.maxcn_mask_fn);
   // *** GET SNV BULK PRIOR ***
   if ( snvEmit.is_set && opts.bulk_fn != NULL ){
     if (opts.bulk_mean)  myClone.allocate_bulk_mean();
@@ -154,28 +156,23 @@ int main (int argc, const char * argv[]){
   }
   //*** GET JUMP PROBABILITY TRACKS and COLLAPSE TO EVENTS***
   get_jump_probability( &myClone, opts);
-  // *** GET TOTAL AND MAXIMUM COPYNUMBER TRACKS***  
-  if ( (snvEmit.is_set || bafEmit.is_set) && opts.cn_fn != NULL ){
-    int global_max = 0;
-    if (snvEmit.is_set) global_max = max( global_max, get_phi( opts.cn_fn,  &snvEmit));
-    if (bafEmit.is_set) global_max = max( global_max, get_phi( opts.cn_fn,  &bafEmit));
-    printf("Found maximum copynumber %i in %s. Will be used as --maxcn.\n\n",
-	   global_max, opts.cn_fn);
-    myClone.maxcn = global_max;//change the maxcn for myClone, fixed hereafter!
-  }
-  else if ( bafEmit.is_set && cnaEmit.is_set==0){
-    bafEmit.cnmax_seen.clear();
-    for (int s=0; s<bafEmit.nSamples; s++){//all chromosomes are normal???
-      bafEmit.cnmax_seen.insert( myClone.normal_copy[bafEmit.chr[s]] );
+  if ( snvEmit.is_set){
+    // *** GET TOTAL MEAN COPYNUMBER TRACKS ***  
+    if( opts.mntcn_fn != NULL ){
+      get_mean_tcn( opts.mntcn_fn, &myClone, &snvEmit);
+    } 
+    // *** GET AVAILABLE COPYNUMBER TRACKS ***  
+    if ( snvEmit.is_set && opts.avcn_fn != NULL ){
+      get_avail_cn( opts.avcn_fn, &myClone, &snvEmit);
     }
-    //bafEmit.cnmax_seen.insert( myClone.maxcn );//or allow all states???
-  }
-  else if ( snvEmit.is_set && cnaEmit.is_set==0){
-    snvEmit.cnmax_seen.clear();
-    for (int s=0; s<snvEmit.nSamples; s++){//all chromosomes are normal???
+    //check whether this is still needed
+    /*if ( cnaEmit.is_set==0){
+      snvEmit.cnmax_seen.clear();
+      for (int s=0; s<snvEmit.nSamples; s++){//all chromosomes are normal???
       snvEmit.cnmax_seen.insert(myClone.normal_copy[snvEmit.chr[s]]);
-    }
-    //snvEmit.cnmax_seen.insert( myClone.maxcn );//or allow all states???
+      }
+      }
+    */
   }
   //*** GET READ DEPTH BIAS FIELD ***
   if (cnaEmit.is_set && opts.bias_fn != NULL){
@@ -439,8 +436,11 @@ void get_opts( int argc, const char ** argv, cmdl_opts& opts){
     else if ( opt_switch.compare("--bulk-updates") == 0 ){
       opts.bulk_updates = atoi(argv[opt_idx]);
     }
-    else if ( opt_switch.compare("--copynumber") == 0){
-      opts.cn_fn = argv[opt_idx];
+    else if ( opt_switch.compare("--mean-tcn") == 0){
+      opts.mntcn_fn = argv[opt_idx];
+    }
+    else if ( opt_switch.compare("--avail-cn") == 0){
+      opts.avcn_fn = argv[opt_idx];
     }
     else if ( opt_switch.compare("--bias") == 0){
       opts.bias_fn = argv[opt_idx];
@@ -565,12 +565,13 @@ void default_opts(cmdl_opts& opts){
   opts.cna_fn    = NULL;
   opts.baf_fn    = NULL;
   opts.snv_fn    = NULL;
-  opts.bulk_fn    = NULL;
-  opts.clones_fn  = NULL;
-  opts.bias_fn    = NULL;
-  opts.cn_fn      = NULL;
-  opts.chr_fn     = NULL;
-  opts.purity_fn  = NULL;
+  opts.bulk_fn   = NULL;
+  opts.clones_fn = NULL;
+  opts.bias_fn   = NULL;
+  opts.mntcn_fn  = NULL;
+  opts.avcn_fn   = NULL;
+  opts.chr_fn    = NULL;
+  opts.purity_fn = NULL;
   opts.cna_jumps_fn   = NULL;
   opts.baf_jumps_fn   = NULL;
   opts.snv_jumps_fn   = NULL;
@@ -599,10 +600,10 @@ void default_opts(cmdl_opts& opts){
   opts.maxcn    = 4;
   opts.min_occ  = 0.01;
   opts.min_jump = 0.01;
-  opts.print_all = 1;
-  opts.bulk_mean=0;
-  opts.bulk_prior=0;
-  opts.bulk_updates=0;
+  opts.print_all = 0;
+  opts.bulk_mean    = 0;
+  opts.bulk_prior   = 0;
+  opts.bulk_updates = 0;
   opts.restarts = 10;
   opts.learn_priors = 1;
   opts.mass_gauging = 1;
@@ -634,24 +635,20 @@ void test_opts(cmdl_opts& opts){
       exit(1);
     }
   }
-  if (opts.cna_fn != NULL && opts.cn_fn != NULL){
-    cout<<"--copynumber [file] cannot be used with --cna [file].\n";
+  if (opts.cna_fn != NULL && (opts.mntcn_fn != NULL || opts.avcn_fn != NULL )){
+    cout<<"--mean-tcn [file] and --avail-cn [file] cannot be used with --cna [file].\n";
     exit(1);
   }
   if ( opts.cna_fn != NULL && opts.cna_jump < 0.0 && opts.cna_jumps_fn == NULL ){
     cout<<"With --cna [file], --cna-jump [double] or --cna-jumps [file] must be given.\n";
     exit(1);
   }
-  /*if ( opts.cna_fn == NULL && opts.baf_fn != NULL && opts.baf_jump < 0.0 && opts.baf_jumps_fn == NULL ){
-    cout<<"With --baf [file] (no CNA), --baf-jump [double] or --baf-jumps [file] must be given.\n";
-    exit(1);
-    }*/
   if (opts.bulk_fix == 0.0 && opts.snv_rnd == 0.0){
     opts.snv_rnd = 1.0e-9;
   }
   //some settings...
   if (opts.clones_fn != NULL){
-    opts.nmax = 100;
+    opts.nmax = 100;//THIS CAUSES A BUG! FIX IT!
   }
   if (opts.force > 0){
     opts.nmax = opts.force;
@@ -659,89 +656,9 @@ void test_opts(cmdl_opts& opts){
 }
 
 void print_opts(){
-  cout<<"# All cloneHD command line options to be used in shell scripts"<<endl;
-  cout<<"# the different data files"<<endl;
-  cout<<"cna=[file]"<<endl;
-  cout<<"baf=[file]"<<endl;
-  cout<<"snv=[file]"<<endl;
-  cout<<"# fixed clone frequencies"<<endl;
-  cout<<"clones=[file]"<<endl;
-  cout<<"# min. purity per sample"<<endl;
-  cout<<"purity=[file]"<<endl;
-  cout<<"# total and max copynumber tracks (from cloneHD with CNA)"<<endl;
-  cout<<"copynumber=[file]"<<endl;
-  cout<<"# the read depth bias field (from filterHD)"<<endl;
-  cout<<"bias=[file]"<<endl;
-  cout<<"# the prefix for all output files"<<endl;
-  cout<<"pre=[string:\"./out\"]"<<endl;
-  cout<<"# a list of chromosomes and their normal copynumber"<<endl;
-  cout<<"chr=[file]"<<endl;
-  cout<<"# the resolution precomputed emission probabilities"<<endl;
-  cout<<"grid=[int:300]"<<endl;
-  cout<<"# random seed"<<endl;
-  cout<<"seed=[int:F(time(0))]"<<endl;
-  cout<<"# max copynumber to consider"<<endl;
-  cout<<"maxcn=[int:4]"<<endl;
-  cout<<"# max no. clones to consider"<<endl;
-  cout<<"nmax=[int:3]"<<endl;
-  cout<<"# force the number of clones"<<endl;
-  cout<<"force=[int]"<<endl;
-  cout<<"# no. independent inference trials (the best is used for output)"<<endl;
-  cout<<"trials=[int:1]"<<endl;
-  cout<<"# no. restarts of local optimization (from perturbed best solution), if applicable"<<endl;
-  cout<<"restarts=[int:10]"<<endl;
-  cout<<"# random emission rates"<<endl;
-  cout<<"cnarnd=[double:0]"<<endl;
-  cout<<"bafrnd=[double:0]"<<endl;
-  cout<<"snvrnd=[double:0]"<<endl;
-  cout<<"# SNV false positive frequency and false positive rate"<<endl;
-  cout<<"snverr=[double:0]"<<endl;
-  cout<<"snvfpr=[double:1.0e-4]"<<endl;
-  cout<<"# jump probabilities per base"<<endl;
-  cout<<"cnajump=[double]"<<endl;
-  cout<<"bafjump=[double]"<<endl;
-  cout<<"snvjump=[double]"<<endl;
-  cout<<"# posterior jump probability track files (from filterHD)"<<endl;
-  cout<<"cnajumps=[file]"<<endl;
-  cout<<"bafjumps=[file]"<<endl;
-  cout<<"snvjumps=[file]"<<endl;
-  cout<<"# dispersion shape parameters"<<endl;
-  cout<<"cnashape=[double:inf]"<<endl;
-  cout<<"bafshape=[double:inf]"<<endl;
-  cout<<"snvshape=[double:inf]"<<endl;
-  cout<<"# penalty term for BAF and SNV priors"<<endl;
-  cout<<"bafpen=[double:0.5]"<<endl;
-  cout<<"snvpen=[double:0.01]"<<endl;
-  cout<<"# min occupancy of level to use for mass gauging"<<endl;
-  cout<<"minocc=[double:0.01]"<<endl;
-  cout<<"# min jump probability for fuzzy segmentation"<<endl;
-  cout<<"minjump=[double:0.01]"<<endl;
-  cout<<"# whether to learn SNV priors"<<endl;
-  cout<<"learnpriors=[0/1:0]"<<endl;
-  cout<<"# whether to do mass-gauging"<<endl;
-  cout<<"massgauging=[0/1:1]"<<endl;  
-  /*
-  cout<<"# bulk allele frequency track (from filterHD, mean or full distribution)"<<endl;
-  cout<<"bulk-prior=[file]"<<endl;
-  cout<<"bulk-mean=[file]"<<endl;
-  cout<<"# fixed allele frequency in the bulk"<<endl;
-  cout<<"bulkfix=[double:0]"<<endl;
-  cout<<"# fixed allele frequency in the bulk"<<endl;
-  cout<<"bulkupdates=[int:0]"<<endl;
-  //cout<<"# diffusion constant for the bulk allele frequency profile"<<endl;
-  //cout<<"bulksigma=[double]"<<endl<<endl;
-  */  
-  cout<<endl<<"cmd=\"./build/cloneHD --cna cna --snv $snv --baf $baf --pre $pre --clones $clones --purity $purity --chr $chr ";
-  cout<<"--bias $bias --copynumber $copynumber --grid $grid --seed $seed --trials $trials ";
-  cout<<"--restarts $restarts --nmax $nmax --force $force --maxcn $maxcn ";
-  cout<<"--cna-jump $cnajump --baf-jump $bafjump --snv-jump $snvjump ";
-  cout<<"--cna-jumps $cnajumps --baf-jumps $bafjumps --snv-jumps $snvjumps ";
-  cout<<"--cna-rnd $cnarnd --baf-rnd $bafrnd --snv-rnd $snvrnd --snv-err $snverr --snv-fpr $snvfpr";
-  cout<<"--cna-shape $cnashape --baf-shape $bafshape --snv-shape $snvshape --baf-pen $bafpen --snv-pen $snvpen";
-  cout<<"--min-occ $minocc --min-jump $minjump --learn-priors $learnpriors --mass-gauging $massgauging";
-  //cout<<"--bulk-prior $bulkprior --bulk-mean $bulkmean --bulk-fix $bulkfix --bulk-sigma $bulksigma --bulk-updates $bulkupdates ";
-  cout<<"\""<<endl<<"echo $cmd"<<endl;
-  cout<<"$cmd"<<endl;
+  cout<<endl<<"./build/cloneHD --cna [file] --snv [file] --baf [file] --pre [string:./out] --clones [file] --purity $purity --chr [file] --bias [file] --mean-tcn [file] --avail-cn [file] --grid [int:300] --seed [int:time(0)] --trials [int:1] --restarts [int:10] --nmax [int:3] --force [int] --maxcn [int:4] --maxcn-mask [file] --cna-jump [double] --baf-jump [double] --snv-jump [double] --cna-jumps [files] --baf-jumps [file] --snv-jumps [file] --cna-rnd [double:0] --baf-rnd [double:0] --snv-rnd [double:0] --snv-err [double] --snv-fpr [double] --cna-shape [double:inf] --baf-shape [double:inf] --snv-shape [double:inf] --baf-pen [double:1.0] --snv-pen [double:0.01] --min-occ [double:0.01] --min-jump [double:0.01] --learn-priors [0/1:0] --mass-gauging [0/1:1] ";
+  //cout<<"--bulk-prior [file] --bulk-mean [file] --bulk-fix [double:0] --bulk-sigma [double] --bulk-updates [int:0]";
+  cout<<endl;
   exit(0);
 }
 
@@ -798,20 +715,21 @@ void print_posterior( FILE * post_fp, Clone * myClone, Emission * myEmit, int s,
   }
 }
 
-void print_phi( FILE * phi_fp, Clone * myClone, Emission * myEmit, int s, cmdl_opts& opts){
+void print_mean_tcn( FILE * mntcn_fp, Clone * myClone, Emission * myEmit, int s, cmdl_opts& opts){
   double ncn = double(myClone->normal_copy[myEmit->chr[s]]);
   for (int evt=0; evt < myEmit->nEvents[s]; evt++){   
     int first = myEmit->idx_of_event[s][evt];     
     int last  = (evt < myEmit->nEvents[s]-1) ?  myEmit->idx_of_event[s][evt+1] - 1 : myEmit->nSites[s]-1; 
     if( opts.print_all || myEmit->coarse_grained == 0 ){
       for (int idx=first; idx<=last; idx++){
-      	fprintf( phi_fp, "%i %6i", myEmit->chr[s], myEmit->loci[s][idx]);
+      	fprintf( mntcn_fp, "%i %6i", myEmit->chr[s], myEmit->loci[s][idx]);
 	for (int t=0; t<myEmit->nTimes; t++){
-	  double tcn = (myEmit->phi==NULL) ? ncn : myEmit->phi[t][s][evt];
-	  fprintf( phi_fp, " %.3f", tcn);
+	  double tcn = (myEmit->mntcn==NULL) ? ncn : myEmit->mntcn[t][s][evt];
+	  fprintf( mntcn_fp, " %.3f", tcn);
 	}
-	int mcn = (myEmit->cnmax==NULL) ? int(ncn) : myEmit->cnmax[s][evt];//???
-	fprintf( phi_fp, " %i\n", mcn);
+	//int mcn = (myEmit->cnmax==NULL) ? int(ncn) : myEmit->cnmax[s][evt];//???
+	//fprintf( phi_fp, " %i\n", mcn);
+	fprintf( phi_fp, "\n");
       }
     }
     else{
@@ -819,11 +737,15 @@ void print_phi( FILE * phi_fp, Clone * myClone, Emission * myEmit, int s, cmdl_o
 	       myEmit->loci[s][first], last-first+1, myEmit->loci[s][last]
 	       );
       for (int t=0; t<myEmit->nTimes; t++){
-	double tcn = (myEmit->phi==NULL) ? ncn : myEmit->phi[t][s][evt];
-	fprintf( phi_fp, " %.3f", tcn);
+	double tcn = (myEmit->mntcn==NULL) ? ncn : myEmit->mntcn[t][s][evt];
+	fprintf( mntcn_fp, " %.3f", tcn);
       }
-      int mcn = (myEmit->cnmax==NULL) ? int(ncn) : myEmit->cnmax[s][evt];//???
-      fprintf( phi_fp, " %i\n", mcn);
+      //int mcn = (myEmit->cnmax==NULL) ? int(ncn) : myEmit->cnmax[s][evt];//???
+      //fprintf( phi_fp, " %i\n", mcn);
+      fprintf( phi_fp, "\n");
     }
   }
 }
+
+//TODO
+//void print_avail_cn( FILE * avcn_fp, Clone * myClone, int s, cmdl_opts& opts){}