+++ emission.h
@@ -82,7 +82,7 @@ public:
   double ** bias;
   double ** log_bias;
   void allocate_bias();
-  void allocate_phi();
+  void allocate_mean_tcn();
   void allocate_cnmax();
   int total_loci, total_events;
   unsigned int total_dist;
@@ -96,9 +96,9 @@ public:
   void coarse_grain_jumps( int sample, double plow, int range);
   double ** pjump;
   void set_pjump(double jump);
-  double *** phi;//total copy number
-  int ** cnmax; //maximum copynumber (across all clones) per site
-  std::set<int> cnmax_seen;
+  double *** mean_tcn;//mean total copy number
+  int ** cnmax;//maximum copynumber (across all clones) per site//???
+  std::set<int> cnmax_seen;//???
   void init_events();
   int * nEvents;
   unsigned int ** Event_of_idx;  // map from idx to cnv-event